feat(console): float carrier job stream above Map panels
The Operations Map only showed a job-count badge on each panel, while Helm streamed in-progress carrier jobs in a dock over the terminal. Extract the shared job-line list into a reusable component and render it on the Map as a dock floating above each panel in canvas space, so it tracks the panel on pan/zoom and never covers the terminal output. Selecting a line opens the same full carrier stream overlay as Helm.
